Revert Greater/Less than intval calls
The change breaks comparing things like dates and I've been unable to
find a test that reproduces the error that caused the change to be added
in the first place. If tests can be added that show why the cast to ints
is needed then they can be added along with the revert of this revert.